requirements.txt
absl-py==2.1.0
accelerate==0.30.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.63.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
kiwisolver==1.4.5
lightning-utilities==0.11.2
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
multidict==6.0.5
natsort==8.4.0
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
Pillow==9.5.0
pip==23.3.1
protobuf==3.20.2
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==1.9.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scipy==1.10.1
sentry-sdk==2.0.1
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1

timm==0.9.16
tokenizers==0.19.1
torchmetrics==1.3.2
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
urllib3==2.2.1
wandb==0.16.6
Werkzeug==3.0.3
wheel==0.43.0
yarl==1.9.4
zipp==3.18.1
pygame==2.6.1
Augmentor==0.2.12
webdriver_manager==4.0.2
selenium==4.32.0
tensorboard==2.9.0
ultralytics==8.3.150
Levenshtein==0.27.1

-f https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html
paddlepaddle-gpu==2.6.1.post120
paddleocr==2.10.0

pytorch_msssim
piq
google-cloud-vision
fpdf
openai